+++ examples/carplay-web-app/src/worker/render/lib/utils.ts
